Cannot delete user posts

+++ CaraLivro/CaraLivro/ViewControllers/UserProfileViewController.swift
@@ -227,10 +227,17 @@ final class UserProfileViewController: UIViewController, Storyboarded, MoreOptio
     public func presentUIAlert(postID: Int, postOwnerID: Int) {
         let alert = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
         self.present(alert, animated: true, completion: nil)
-        
-        alert.addAction(UIAlertAction(title: "Apagar", style: .destructive, handler: { action in
-            self.presenter?.deletePost(postID: postID)
-        }))
+        if (presenter?.currentUser?.idUserProfile == currentUserInUse?.idUserProfile) {
+            alert.addAction(UIAlertAction(title: "Apagar", style: .destructive, handler: { action in
+                self.presenter?.deletePost(postID: postID)
+            }))
+        } else {
+            alert.addAction(UIAlertAction(title: "Denunciar", style: .destructive, handler: { action in
+                let alert = UIAlertController(title: "Denunciar", message: "Obrigado pela denúncia! :)", preferredStyle: UIAlertControllerStyle.alert)
+                alert.addAction(UIAlertAction(title: "Ok", style: UIAlertActionStyle.default, handler: nil))
+                self.present(alert, animated: true, completion: nil)
+            }))
+        }
         alert.addAction(UIAlertAction(title: "Cancelar", style: .cancel, handler: nil))
     }
